refactor(#3269): extract loop-analysis.ts pure predicates from loops.ts (#3182)
Verbatim move of 14 static-analysis predicates + isAssignmentOperator DRY.
emit-identity IDENTICAL, tsc 0.

✓
